+++ python/acl_anthology/collections/event.py
@@ -1,5 +1,5 @@
 # Copyright 2022 Matt Post <post@cs.jhu.edu>
-# Copyright 2023-2025 Marcel Bollmann <marcel@bollmann.me>
+# Copyright 2023-2026 Marcel Bollmann <marcel@bollmann.me>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.