+++ python/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Files (such as paper PDFs) can now be downloaded from their remote URLs via `.download()`.
+- Added `.get_namespec_for(Person)` on papers and volumes, to more easily find the NameSpecification referring to a given Person.
+- Added `Person.namespecs()` to iterate over all NameSpecifications referring to this Person.
 
 ### Changed